fix aws bug
Former-commit-id: 102450dbabb7c245466a5097305f4659b543a6b9
Former-commit-id: 0855a47add6623006e7d1f250978a18320c909ba

+++ emirt/volume_util.py
@@ -80,7 +80,7 @@ def norm(vol):
 	'''Normalizes the input volume to have values between 0 and 1
 	(achieved by factor normalization to the max)'''
 
-	vol = vol - np.min(vol)
+	vol = vol - np.min(vol.astype('float32'))
 	vol = vol / np.max(vol)
 
 	return vol
